Split SkeletonBuilder out into Minutia and SkeletonBuilderRidge.
